Distance from plate boundary

Distance of data points from the nearest plate boundary in degree

Usage

distance_from_pb(x, ep, pb, tangential = FALSE)

Arguments

x, pb

sf objects of the data points and the plate boundary geometries in the geographical coordinate system

ep

data.frame of the geographical coordinates of the Euler pole (lat, lon)

tangential

Logical. TRUE for tangential boundaries and FALSE (the default) for inward and outward boundaries.

Value

Numeric vector of the (great circle) distances in degree

Examples

data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("PB2002")
san_andreas <- subset(PB2002, PB2002$PlateA %in% c("NA" , "PA") & PB2002$PlateB %in% c("NA", "PA"))

data("wsm2016")
california <- subset(
wsm2016,
wsm2016$lat >= 23 & wsm2016$lat <= 40 &
  wsm2016$lon >= -126 & wsm2016$lon <= -108
  )
california <- sf::st_set_crs(sf::st_as_sf(california, coords = c("lon", "lat")), "WGS84")

distance_from_pb(x = california, ep = na_pa, pb = san_andreas, tangential = TRUE)
#>    [1] 1.398009e-01 8.983626e-02 1.794734e-01 1.541420e-01 9.999110e-04
#>    [6] 9.028785e-03 1.092096e-01 1.445477e-01 6.113996e-02 5.033963e-02
#>   [11] 8.316812e-02 1.341720e-01 1.610081e-01 2.747857e-01 2.736191e-01
#>   [16] 4.344864e-02 1.262250e-01 8.235719e-02 1.393882e-02 9.898493e-03
#>   [21] 8.304070e-02 1.210834e-01 1.078401e-01 1.725346e-02 4.173973e-02
#>   [26] 2.511176e-01 5.080613e-02 1.823931e-01 6.124010e-02 6.486728e-02
#>   [31] 3.616363e-02 2.126209e-02 6.043115e-02 1.081778e-01 1.632153e-01
#>   [36] 6.493690e-02 6.493690e-02 4.163589e-02 2.053917e-02 2.634297e-02
#>   [41] 1.934664e-01 2.022201e-01 5.642771e-02 1.808823e-02 4.275987e-02
#>   [46] 2.027991e-01 2.027991e-01 1.460455e-02 1.571269e-01 2.405830e-01
#>   [51] 6.014656e-02 6.504090e-02 7.174077e-02 3.153326e-01 1.979670e-01
#>   [56] 2.349741e-01 2.584352e-01 1.936184e-01 1.936184e-01 3.299719e-02
#>   [61] 8.957626e-02 2.636975e-02 1.650956e-01 1.083776e-01 8.082132e-02
#>   [66] 8.937405e-02 8.937405e-02 1.350215e-01 1.082448e-01 5.224751e-02
#>   [71] 1.191793e-01 3.516986e-02 5.622517e-02 4.890131e-02 1.206118e-01
#>   [76] 1.387269e-01 3.476377e-02 1.426976e-02 1.105361e-01 1.257085e-01
#>   [81] 6.273036e-02 9.954496e-02 1.492189e-01 4.163589e-02 2.392887e-01
#>   [86] 2.053917e-02 3.391044e-02 2.634297e-02 1.430190e-01 8.867094e-03
#>   [91] 1.365914e-01 7.593147e-02 1.934664e-01 1.393152e-01 3.592873e-02
#>   [96] 2.022201e-01 1.346979e-01 3.321490e-01 1.808823e-02 1.039815e-01
#>  [101] 1.147222e-01 4.275987e-02 1.183572e-01 3.775795e-02 1.460455e-02
#>  [106] 2.268384e-01 1.549940e-01 1.889528e-01 2.405830e-01 7.174077e-02
#>  [111] 2.044705e-01 3.153326e-01 5.796875e-02 6.941066e-02 2.878408e-01
#>  [116] 1.673559e-01 1.954145e-01 1.979670e-01 3.073768e-01 2.349741e-01
#>  [121] 8.650141e-03 1.701747e-01 2.278684e-01 1.208714e-01 1.174719e-01
#>  [126] 1.905235e-01 9.088731e-02 2.132500e-01 8.751865e-03 1.147186e-01
#>  [131] 1.112162e-01 1.875892e-01 2.254503e-01 2.797155e-01 2.584352e-01
#>  [136] 2.284436e-01 1.658544e-01 1.373142e-01 2.445293e-01 7.298305e-03
#>  [141] 2.192310e-01 1.640592e-01 1.101834e-01 3.529565e-02 4.889697e-02
#>  [146] 2.689728e-01 3.591046e-02 8.957626e-02 1.048084e-01 2.636975e-02
#>  [151] 2.206381e-01 2.079413e-01 2.873371e-01 1.183683e-01 3.312624e-02
#>  [156] 3.995959e-02 4.865283e-02 9.956248e-02 2.565878e-02 2.334225e-02
#>  [161] 6.072271e-02 7.161099e-03 1.431937e-02 2.328935e-03 4.480961e-02
#>  [166] 2.930291e-02 1.086338e-01 5.148571e-02 2.140430e-02 2.320754e-02
#>  [171] 8.398762e-02 1.350215e-01 8.030981e-02 1.925845e-01 4.375644e-02
#>  [176] 1.617765e-01 4.296239e-02 1.995755e-01 8.809786e-02 1.707947e-02
#>  [181] 4.747694e-02 4.747694e-02 2.934743e-03 2.555783e-01 1.913396e-01
#>  [186] 8.342801e-02 1.899293e-01 8.517184e-02 1.346756e-01 5.223406e-03
#>  [191] 6.422826e-02 1.059704e-01 1.188531e-02 2.630372e-02 1.261260e-02
#>  [196] 2.007509e-02 1.325666e-02 9.175127e-02 9.719802e-02 3.363167e-02
#>  [201] 1.797994e-01 9.352509e-03 9.221690e-02 1.873795e-01 1.215868e-01
#>  [206] 9.172045e-03 6.900364e-02 1.184749e-02 4.435764e-03 5.083273e-02
#>  [211] 1.325207e-01 8.734013e-02 6.331273e-02 1.990526e-02 1.279608e-01
#>  [216] 6.130915e-02 1.704075e-01 4.759270e-02 4.915235e-03 2.498014e-02
#>  [221] 3.024325e-02 9.035070e-02 7.392102e-02 1.167330e-01 1.511820e-01
#>  [226] 1.013905e-01 1.578942e-01 1.473276e-01 1.573707e-01 1.037705e-01
#>  [231] 1.012371e-01 6.257910e-02 1.181127e-01 1.953547e-01 7.868573e-02
#>  [236] 1.017056e-01 5.622586e-02 1.166508e-01 1.278986e-01 1.922036e-01
#>  [241] 2.294563e-02 9.477963e-02 2.825455e-01 1.954486e-01 9.355166e-03
#>  [246] 1.603080e-01 3.367552e-02 1.398661e-02 5.648932e-02 1.746155e-01
#>  [251] 1.550846e-01 1.163462e-01 2.962778e-01 2.416554e-02 5.612145e-02
#>  [256] 1.648841e-02 5.578586e-02 5.713092e-02 8.486526e-02 8.707790e-02
#>  [261] 5.512512e-02 5.518753e-02 6.112657e-02 1.502833e-02 9.642999e-04
#>  [266] 2.793019e-02 5.019997e-02 1.176201e-02 8.170978e-03 5.384525e-02
#>  [271] 6.856079e-02 5.094569e-02 2.264413e-01 1.662862e-01 1.129247e-01
#>  [276] 7.055630e-02 4.885607e-02 1.410906e-01 5.497552e-02 8.832135e-02
#>  [281] 1.702383e-02 7.115180e-02 1.158988e-01 4.318790e-02 1.192186e-01
#>  [286] 2.470230e-02 7.362074e-02 1.105207e-01 7.717138e-02 2.792143e-02
#>  [291] 2.828805e-02 8.436409e-02 6.965210e-03 1.894556e-02 3.369715e-02
#>  [296] 4.984435e-02 5.808858e-02 6.275698e-02 1.087260e-01 1.078401e-01
#>  [301] 3.400199e-02 2.145204e-01 2.315594e-01 1.745268e-01 1.936355e-01
#>  [306] 1.141717e-01 2.436728e-02 5.077925e-02 1.010947e-01 4.416400e-02
#>  [311] 9.393612e-02 7.763344e-02 5.759110e-02 2.958213e-02 5.555439e-02
#>  [316] 2.202563e-01 8.171473e-02 1.306032e-01 8.802458e-02 1.690811e-03
#>  [321] 7.778575e-02 1.535205e-01 2.415644e-02 6.922181e-02 2.264994e-02
#>  [326] 1.511718e-01 9.154245e-02 1.005398e-02 8.530286e-02 1.761694e-02
#>  [331] 1.062993e-01 3.337371e-02 1.636512e-02 1.218306e-01 3.319469e-02
#>  [336] 7.780278e-02 7.683450e-03 6.266187e-02 9.492074e-03 2.906036e-02
#>  [341] 9.865591e-02 2.094927e-02 5.440155e-02 1.076626e-02 9.506607e-02
#>  [346] 7.594114e-02 1.932310e-02 9.365766e-02 6.515991e-03 3.013882e-02
#>  [351] 3.013882e-02 2.226246e-02 1.261260e-02 3.324944e-02 9.162350e-02
#>  [356] 5.915576e-02 2.951863e-02 2.951863e-02 4.802754e-02 6.956702e-02
#>  [361] 2.834581e-02 2.834581e-02 6.013979e-02 4.364054e-02 7.285564e-02
#>  [366] 5.294782e-02 3.551898e-02 6.035169e-03 4.013306e-03 4.612656e-02
#>  [371] 7.321100e-03 3.300797e-03 2.391991e-02 4.909627e-02 1.799025e-02
#>  [376] 3.427887e-02 4.861386e-02 8.359995e-02 6.138001e-02 1.473276e-01
#>  [381] 1.129553e-01 5.095829e-02 1.963831e-01 9.935566e-02 1.249949e-01
#>  [386] 2.989552e-02 2.458894e-03 7.455507e-02 1.617765e-01 1.642958e-01
#>  [391] 2.641418e-02 1.697312e-01 1.252582e-01 3.259593e-01 3.292089e-03
#>  [396] 1.684441e-01 5.097590e-02 5.009073e-02 5.871438e-02 1.822526e-02
#>  [401] 8.495725e-02 3.636800e-02 1.078641e-01 3.564860e-02 7.411359e-02
#>  [406] 1.346904e-01 1.765789e-01 9.482820e-02 1.368957e-01 1.674910e-01
#>  [411] 3.530260e-02 1.540033e-01 1.611355e-01 1.940196e-01 3.514810e-02
#>  [416] 6.103272e-03 7.721295e-03 3.791694e-03 2.634255e-01 2.440753e-01
#>  [421] 6.797848e-02 8.739442e-02 1.142596e-01 9.792338e-03 4.477683e-02
#>  [426] 7.199217e-02 7.068642e-02 1.376746e-02 4.536867e-02 6.794116e-02
#>  [431] 1.183105e-01 8.243520e-03 5.345257e-02 3.736701e-02 4.751513e-02
#>  [436] 1.032344e-01 6.891547e-02 7.957020e-02 4.727717e-03 6.144754e-02
#>  [441] 1.938786e-02 2.047678e-02 4.601064e-03 2.174379e-03 3.735918e-03
#>  [446] 1.508306e-01 9.818673e-02 3.221538e-03 1.092018e-04 2.541504e-03
#>  [451] 3.568216e-02 9.359759e-04 7.078743e-03 2.550356e-02 2.428879e-02
#>  [456] 8.721751e-03 6.530113e-03 1.700450e-02 1.081205e-02 2.051941e-02
#>  [461] 1.381532e-02 1.768769e-03 1.703735e-02 5.088971e-02 5.046602e-02
#>  [466] 1.203874e-02 1.604923e-02 3.860338e-02 2.291195e-02 5.428071e-02
#>  [471] 1.664643e-01 1.757291e-01 6.301442e-02 4.698121e-02 4.330562e-02
#>  [476] 4.905674e-02 8.647315e-02 1.414103e-01 1.256613e-01 6.606142e-04
#>  [481] 3.191884e-02 1.212677e-01 6.694328e-02 1.806483e-01 8.062754e-02
#>  [486] 1.680537e-02 1.998334e-02 4.305803e-03 1.105434e-02 1.694945e-01
#>  [491] 1.738891e-01 1.676354e-01 1.073504e-01 9.200225e-02 7.564509e-02
#>  [496] 6.027811e-02 4.423029e-02 1.301835e-01 1.470741e-01 1.002522e-01
#>  [501] 7.501463e-02 9.118590e-02 1.703174e-01 1.911715e-01 9.612270e-02
#>  [506] 3.123447e-02 9.496822e-02 1.595124e-01 1.871483e-01 1.399217e-01
#>  [511] 3.563093e-02 5.182824e-02 1.289059e-02 1.687834e-01 7.025491e-02
#>  [516] 8.562320e-02 1.016771e-01 1.180676e-01 1.495446e-01 1.652881e-01
#>  [521] 1.813868e-01 5.171617e-02 3.585831e-02 1.965285e-02 4.106876e-03
#>  [526] 1.244744e-02 5.320862e-03 1.855179e-01 1.661955e-02 4.880986e-02
#>  [531] 8.103705e-02 2.133707e-03 1.391904e-02 4.570202e-02 6.178161e-02
#>  [536] 7.719398e-02 9.362008e-02 1.090514e-01 9.122411e-02 7.567136e-02
#>  [541] 1.890502e-01 1.063544e-01 7.461232e-02 5.855167e-02 4.215420e-02
#>  [546] 2.675390e-02 1.008271e-01 1.173161e-01 1.307549e-01 1.145069e-01
#>  [551] 4.138099e-02 1.377231e-01 1.471804e-01 1.311120e-01 1.147066e-01
#>  [556] 6.101476e-02 7.719960e-02 1.540347e-01 1.175717e-01 5.221308e-02
#>  [561] 1.873482e-01 1.719348e-01 1.086885e-02 4.978307e-03 2.118007e-02
#>  [566] 3.739058e-02 1.663716e-01 1.037907e-01 9.007909e-02 1.558900e-01
#>  [571] 9.970895e-02 6.766351e-02 3.523541e-02 3.116113e-03 1.312922e-02
#>  [576] 9.447937e-02 1.107779e-01 1.264133e-01 1.562556e-01 1.397167e-01
#>  [581] 1.241775e-01 5.888803e-02 4.297019e-02 2.669925e-02 3.813145e-02
#>  [586] 5.477982e-02 8.743701e-02 1.034405e-01 6.703490e-03 2.235086e-02
#>  [591] 1.086727e-01 1.407647e-01 1.738976e-01 1.874475e-01 1.804251e-01
#>  [596] 1.641822e-01 1.479304e-01 1.313379e-01 9.946656e-02 8.284653e-02
#>  [601] 6.655002e-02 1.616748e-03 4.776630e-02 6.346153e-02 1.107924e-01
#>  [606] 6.314363e-02 1.311189e-01 1.473694e-01 1.639607e-01 1.802291e-01
#>  [611] 1.879316e-01 1.719909e-01 1.393911e-01 1.230781e-01 1.020201e-01
#>  [616] 6.929710e-02 2.908059e-02 6.228087e-02 1.368982e-01 1.204185e-01
#>  [621] 1.042593e-01 8.809113e-02 9.099356e-02 1.072695e-01 1.721165e-01
#>  [626] 1.793378e-01 1.629993e-01 1.158809e-01 1.323868e-01 1.320964e-01
#>  [631] 1.487762e-01 1.651314e-01 1.814954e-01 3.773316e-02 1.817781e-01
#>  [636] 4.265944e-02 1.244626e-01 1.575859e-01 5.174694e-02 1.064711e-01
#>  [641] 6.804321e-02 8.444955e-02 4.262429e-02 9.542095e-03 8.958661e-02
#>  [646] 2.798294e-02 4.441459e-02 1.209914e-02 4.357797e-03 8.615730e-02
#>  [651] 6.969269e-02 5.288288e-02 5.563340e-02 1.466301e-02 1.263341e-01
#>  [656] 1.222920e-01 1.274361e-01 1.442882e-01 2.657228e-02 3.534524e-02
#>  [661] 5.348034e-02 7.000288e-02 8.653439e-02 1.034125e-01 3.113761e-02
#>  [666] 2.400539e-03 3.527971e-02 7.977108e-02 9.737420e-02 4.289094e-02
#>  [671] 1.193223e-01 7.376006e-02 8.551850e-02 6.889291e-02 5.259058e-02
#>  [676] 3.955857e-02 5.650631e-02 7.732975e-02 1.098686e-01 4.881003e-02
#>  [681] 5.140859e-02 7.334529e-02 3.985358e-02 5.981413e-03 1.484868e-02
#>  [686] 3.110325e-02 1.142418e-01 1.157483e-02 5.413602e-03 2.204416e-02
#>  [691] 3.835142e-02 5.536604e-02 7.202458e-02 8.835924e-02 1.054000e-01
#>  [696] 1.220865e-01 7.490441e-02 5.792358e-02 1.851983e-02 1.589832e-02
#>  [701] 3.692709e-02 2.060383e-02 3.572424e-03 3.015128e-02 6.321372e-02
#>  [706] 1.332198e-01 1.241232e-01 5.658273e-02 5.193943e-02 1.664798e-02
#>  [711] 1.133247e-01 9.596069e-02 7.962149e-02 5.850646e-02 6.480891e-02
#>  [716] 1.387014e-01 1.049292e-01 9.800508e-02 2.066731e-01 1.766919e-01
#>  [721] 1.431516e-01 1.218153e-01 1.053461e-01 1.132507e-01 3.608150e-02
#>  [726] 5.104820e-02 1.197014e-01 1.469357e-01 1.297339e-01 8.868424e-02
#>  [731] 1.312818e-01 1.723540e-01 1.287435e-01 6.401088e-02 7.074555e-02
#>  [736] 1.044873e-01 2.080008e-01 1.749814e-01 8.477221e-02 5.710290e-02
#>  [741] 1.308488e-01 6.166243e-02 1.121876e-01 4.244347e-02 1.527841e-02
#>  [746] 3.253788e-02 5.014407e-02 6.742078e-02 8.400531e-02 1.362524e-01
#>  [751] 1.532125e-01 1.705414e-01 1.878789e-01 4.875442e-02 4.212448e-03
#>  [756] 4.229176e-02 5.961096e-02 7.657791e-02 1.112601e-01 1.289539e-01
#>  [761] 1.720230e-02 1.036951e-01 8.771308e-02 1.056517e-01 1.232538e-01
#>  [766] 3.976165e-02 1.090050e-01 4.115680e-02 1.140266e-01 4.738238e-02
#>  [771] 8.228479e-02 8.952599e-02 1.893864e-01 8.718352e-02 7.248063e-02
#>  [776] 8.995319e-02 1.163888e-02 6.342320e-03 2.398617e-02 4.163841e-02
#>  [781] 5.964547e-02 4.464765e-02 2.549608e-02 9.508034e-02 3.057258e-02
#>  [786] 4.841291e-02 6.968586e-02 5.203411e-02 3.437395e-02 1.635880e-02
#>  [791] 6.289005e-04 1.831484e-02 9.313968e-02 7.545427e-02 5.776044e-02
#>  [796] 3.971092e-02 1.579085e-01 1.335738e-01 1.158636e-01 9.814495e-02
#>  [801] 2.325492e-02 9.442624e-02 1.653901e-01 1.662081e-01 8.360792e-02
#>  [806] 1.175084e-02 7.263842e-02 1.659858e-01 1.301495e-01 4.225934e-02
#>  [811] 1.391259e-02 1.314014e-01 5.940449e-02 1.889405e-01 1.428401e-02
#>  [816] 3.904815e-02 9.677361e-02 1.236021e-01 6.454682e-02 3.879069e-02
#>  [821] 3.307013e-02 2.008035e-01 5.615636e-02 4.259068e-02 3.873417e-02
#>  [826] 1.332577e-01 1.047103e-01 7.624761e-02 1.232109e-03 8.442536e-03
#>  [831] 2.083149e-01 2.496994e-02 8.366313e-02 4.876123e-02 3.111310e-02
#>  [836] 3.094188e-02 4.150785e-02 6.761088e-02 2.970457e-02 4.655084e-04
#>  [841] 9.450408e-02 1.020693e-01 2.687965e-02 1.048298e-01 1.264394e-02
#>  [846] 2.451851e-01 8.717588e-02 6.728696e-02 2.079733e-01 2.355556e-01
#>  [851] 7.006280e-02 3.549668e-02 2.735577e-02 4.071030e-02 6.438481e-02
#>  [856] 6.847568e-02 7.006560e-02 6.773703e-02 5.943383e-03 2.164878e-02
#>  [861] 4.213687e-02 6.830571e-02 9.238666e-03 6.103174e-02 1.426226e-02
#>  [866] 7.814855e-02 3.106253e-02 6.154299e-02 6.230740e-02 2.867402e-02
#>  [871] 4.071178e-02 2.164878e-02 6.971811e-02 3.705264e-03 2.015866e-02
#>  [876] 5.697455e-03 2.713362e-02 8.953412e-02 2.124439e-02 4.898805e-02
#>  [881] 2.353667e-02 6.273882e-02 1.477877e-02 1.786962e-02 4.055055e-02
#>  [886] 2.855790e-01 7.200216e-03 1.289129e-02 2.707621e-02 1.720854e-02
#>  [891] 6.149587e-02 9.743709e-02 1.030391e-01 2.967433e-03 1.121212e-01
#>  [896] 1.423907e-01 1.689946e-02 1.902116e-01 3.493479e-02 1.507332e-01
#>  [901] 1.258126e-01 1.116193e-01 1.043930e-01 5.636954e-02 2.708864e-01
#>  [906] 8.867010e-02 3.151242e-02 1.491868e-01 2.736622e-01 2.067042e-02
#>  [911] 2.496492e-01 5.126029e-02 4.053491e-02 1.439569e-01 1.034925e-01
#>  [916] 1.589040e-01 6.696323e-02 1.995567e-02 1.580625e-01 3.773673e-02
#>  [921] 4.139067e-02 4.607607e-02 2.143637e-03 1.029656e-02 2.025342e-02
#>  [926] 3.852358e-03 1.981536e-02 1.806702e-02 3.250967e-02 4.131643e-02
#>  [931] 2.394422e-04 3.699403e-02 1.548234e-02 9.516037e-02 6.616744e-02
#>  [936] 8.187103e-02 2.167836e-02 2.674016e-02 7.483886e-02 5.881179e-02
#>  [941] 1.082206e-01 7.886661e-02 7.630106e-02 1.057105e-01 1.138266e-01
#>  [946] 5.270625e-02 8.776334e-02 1.198813e-01 7.405038e-02 4.840853e-02
#>  [951] 4.055566e-02 9.008663e-03 3.132662e-02 7.001263e-03 5.173000e-04
#>  [956] 1.676691e-02 1.024273e-02 1.360024e-02 1.427066e-02 1.891837e-02
#>  [961] 2.484555e-02 6.761387e-02 3.074967e-02 3.726381e-02 5.513856e-02
#>  [966] 1.921714e-02 8.918030e-02 7.162998e-02 5.789397e-02 6.139480e-02
#>  [971] 8.677756e-03 2.830812e-04 3.428692e-02 6.884736e-02 1.017210e-02
#>  [976] 5.176292e-02 1.035446e-01 4.841526e-03 2.085892e-01 2.085892e-01
#>  [981] 2.085892e-01 4.196567e-02 1.110420e-02 5.932080e-02 9.055378e-02
#>  [986] 5.054810e-02 3.077017e-02 8.304178e-03 1.817654e-02 1.369268e-01
#>  [991] 4.697755e-02 1.272476e-01 2.145647e-02 2.145647e-02 2.123006e-02
#>  [996] 3.546413e-03 6.617855e-03 1.813091e-02 7.247675e-02 5.004094e-02
#> [1001] 1.056345e-02 2.802562e-02 7.423931e-02 1.374567e-02 1.300639e-02
#> [1006] 7.074876e-02 2.187204e-02 5.696615e-03 2.211025e-02 1.595583e-02
#> [1011] 2.950004e-02 1.056060e-02 7.598637e-02 5.253544e-02 2.929237e-02
#> [1016] 7.807458e-03 4.711683e-02 2.073057e-02 5.816190e-03 1.501266e-02
#> [1021] 3.068948e-03 5.223464e-02 1.433252e-01 2.807210e-02 3.195403e-02
#> [1026] 2.769350e-02 3.387214e-02 4.791112e-02 2.442886e-02 1.327481e-01
#> [1031] 2.860528e-01 6.375806e-02 1.756562e-02 7.525194e-02 1.760892e-02
#> [1036] 4.047506e-02 6.135732e-02 1.975893e-02 5.041903e-03 5.072419e-03
#> [1041] 1.127070e-02 3.497060e-02 1.715036e-02 4.328038e-02 4.969277e-02
#> [1046] 2.243047e-02 2.929359e-02 5.812773e-02 2.403255e-02 1.081506e-01
#> [1051] 3.058861e-02 7.529283e-02 4.002256e-02 9.833862e-02 1.300624e-01
#> [1056] 2.181475e-01 3.659423e-02 7.263879e-02 3.852736e-02 1.832854e-02
#> [1061] 2.639557e-01 7.706205e-02 1.105824e-01 1.639523e-02 3.508162e-03
#> [1066] 1.363166e-02 9.523859e-02 2.178375e-02 5.137690e-02 5.211843e-02
#> [1071] 1.356702e-02 6.099358e-02 9.967131e-02 6.578155e-02 6.257515e-02
#> [1076] 4.740306e-02 2.576121e-02 2.398578e-05 4.084473e-02 9.454486e-03
#> [1081] 4.571661e-02 6.809760e-02 1.374290e-02 7.571509e-02 1.017547e-01
#> [1086] 9.232646e-02 6.367777e-02 1.324337e-01 6.813386e-02 6.027224e-02
#> [1091] 7.689826e-03 5.866160e-02 5.866160e-02 3.789260e-02 2.717255e-01
#> [1096] 2.835666e-01 2.153472e-01 3.293899e-01 2.668115e-01 3.204349e-01
#> [1101] 2.122380e-01 2.472088e-01 3.174357e-01 2.870184e-01 2.834777e-01
#> [1106] 2.940908e-01 3.179277e-01 2.826967e-01 2.755104e-01 2.678940e-01
#> [1111] 2.612054e-01 2.798848e-01 2.540275e-01 2.086421e-01 3.059636e-01
#> [1116] 3.050537e-01 3.304024e-01 3.354040e-01 3.022068e-01 2.640952e-01
#> [1121] 2.897993e-01 3.135924e-01 3.256374e-01 2.990531e-01 2.587816e-01
#> [1126] 3.327821e-01 3.318926e-01 3.318926e-01 3.035510e-01 3.068582e-01
#> [1131] 2.764233e-01 2.824317e-01 7.480655e-02 5.948044e-03 5.443122e-02
#> [1136] 5.756765e-02 5.756765e-02 5.758520e-02 5.758520e-02 5.359362e-03
#> [1141] 5.359362e-03 2.737424e-01