Great-circles of Euler pole

Get points on a great circle as defined by the shortest distance between two specified points

Usage

eulerpole_greatcircles(x, gm, n, sf = TRUE)

Arguments

x

data.frame containing coordinates of Euler poles in lat, lon, and rotation angle (optional)

gm

numeric, angle between great circles

n

numeric; number of equally spaced great circles (angle between great circles (number of great circles n = 360 / gm); default = 12

sf

logical. Export object type of the lines. TRUE (the default) simple feature (sf) objects. FALSE for "SpatialLInesDataFrame" object.

Value

An object of class sf or SpatialLinesDataFrame

Author

Tobias Stephan

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to Pacific plate
euler$angle <- euler$rate
eulerpole_greatcircles(euler)
#> Simple feature collection with 13 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 90
#> CRS:           +proj=longlat +datum=WGS84
#> First 10 features:
#>      greatcircle                       geometry
#> -180        -180 LINESTRING (-180 -90, -180 ...
#> -150        -150 LINESTRING (150 -90, 150 -8...
#> -120        -120 LINESTRING (120 -90, 120 -8...
#> -90          -90 LINESTRING (90 -90, 90 -89,...
#> -60          -60 LINESTRING (60 -90, 60 -89,...
#> -30          -30 LINESTRING (30 -90, 30 -89,...
#> 0              0 LINESTRING (0 -90, 0 -89, 0...
#> 30            30 LINESTRING (-30 -90, -30 -8...
#> 60            60 LINESTRING (-60 -90, -60 -8...
#> 90            90 LINESTRING (-90 -90, -90 -8...