Small-circles around Euler pole

Construct a line that has a constant distance to the Euler pole.

Usage

eulerpole_smallcircles(x, sm, sf = TRUE)

Arguments

x

data.frame containing coordinates of Euler pole in lat, lon, and rotation angle (optional)

sm

numeric, angle between small circles (in degree); default: 10

sf

logical. Export object type of the lines. TRUE (the default) simple feature (sf) objects. FALSE for "SpatialLInesDataFrame" object.

Value

An object of class sf or SpatialLinesDataFrame

Details

if angle is given: output additionally gives absolute velocity on small circle (degree/Myr -> km/Myr)

Author

Tobias Stephan

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
euler$angle <- euler$rate
# Pacific plate
eulerpole_smallcircles(euler)
#> Simple feature collection with 19 features and 2 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 90
#> CRS:           +proj=longlat +datum=WGS84
#> First 10 features:
#>    small_circle.small_circle small_circle.abs_vel
#> 0                          0              0.00000
#> 10                        10             15.11688
#> 20                        20             29.77444
#> 30                        30             43.52731
#> 40                        40             55.95764
#> 50                        50             66.68772
#> 60                        60             75.39152
#> 70                        70             81.80459
#> 80                        80             85.73207
#> 90                        90             87.05463
#>                          geometry
#> 0  LINESTRING (-180 -90, -170 ...
#> 10 LINESTRING (-180 -80, -170 ...
#> 20 LINESTRING (-180 -70, -170 ...
#> 30 LINESTRING (-180 -60, -170 ...
#> 40 LINESTRING (-180 -50, -170 ...
#> 50 LINESTRING (-180 -40, -170 ...
#> 60 LINESTRING (-180 -30, -170 ...
#> 70 LINESTRING (-180 -20, -170 ...
#> 80 LINESTRING (-180 -10, -170 ...
#> 90 LINESTRING (-180 0, -170 0,...