Azimuth between two points

Calculate initial bearing (or forward azimuth/direction) to go from point p1 to point p2 following great-circle arc on a sphere.

Usage

get_azimuth(p, q)

Arguments

p

latitude/longitude of start point(s). Can be a vector of two numbers or a matrix of 2 columns (first one is latitude, second is longitude)

q

as above

Value

azimuth in degrees

Details

This formula is for the initial bearing (sometimes referred to as forward azimuth) which if followed in a straight line along a great-circle arc will take you from the start point to the end point: \(\theta = \arctan2 (\sin \Delta\lambda * \cos\psi_2, \cos\psi_1*\sin\psi_1-\sin\psi_1*\cos\psi_2*\cos\Delta\lambda)\) where \(\psi_1, \lambda_1\) is the start point, \(\psi_2\), \(\lambda_2\) the end point (\(\Delta\lambda\) is the difference in longitude)

References

http://www.movable-type.co.uk/scripts/latlong.html

C.F.F. Karney, 2013. Algorithms for geodesics, J. Geodesy 87: 43-55. https://dx.doi.org/10.1007/s00190-012-0578-z

Author

Tobias Stephan

Examples

p <- c(35, 45) # Baghdad
q <- c(35, 135) # Osaka
get_azimuth(p, q)
#> [1] 60.16243