Theoretical Direction of Maximum Horizontal Stress Models the direction of maximum horizontal stress along great circles, small circles, and loxodromes at a given point or points according to the relative plate motion

Theoretical Direction of Maximum Horizontal Stress

Models the direction of maximum horizontal stress along great circles, small circles, and loxodromes at a given point or points according to the relative plate motion

Usage

model_shmax(df, euler)

Arguments

df

data.frame containing the coordinates of the point(s) (lat, lon)

euler

data.frame containing the coordinates of the Euler pole for the plate boundary (lat, lon)

Value

An object of class data.frame

gc

Azimuth of the modeled maximum horizontal following stress great circles

sc

Small circles

ld.cw

Clockwise loxodromes

ld.ccw

Counter-clockwise loxodromes

@seealso misfit_shmax() to compute the deviation of the modeled direction from the observed direction of maximum stress.

References

Wdowinski, S., 1998, A theory of intraplate tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p. 5037-5059, http://dx.doi.org/10.1029/97JB03390.

Author

Tobias Stephan

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
# Pacific plate
point <- data.frame(lat = 45, lon = 20)
model_shmax(point, euler)
#>         gc       sc   ld.ccw    ld.cw
#> 1 54.55267 42.45436 87.45436 177.4544