Distance from plate boundary

Distance of data points from the nearest plate boundary in degree

Usage

distance_from_pb(x, ep, pb, tangential = FALSE)

Arguments

x, pb

sf objects of the data points and the plate boundary geometries in the geographical coordinate system

ep

data.frame of the geographical coordinates of the Euler pole (lat, lon)

tangential

Logical. TRUE for tangential boundaries and FALSE (the default) for inward and outward boundaries.

Value

Numeric vector of the great circle distances in degree

Details

The distance to the plate boundary is the longitudinal or latitudinal difference between the data point and the plate boundary (along the closest latitude or longitude) for inward/outward or tangential plate boundaries, respectively.

Examples

data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("PB2002")
san_andreas <- subset(PB2002, PB2002$PlateA %in% c("NA" , "PA") & PB2002$PlateB %in% c("NA", "PA"))

data("wsm2016")
california <- subset(
wsm2016,
wsm2016$lat >= 23 & wsm2016$lat <= 40 &
  wsm2016$lon >= -126 & wsm2016$lon <= -108
  )
california <- sf::st_set_crs(sf::st_as_sf(california, coords = c("lon", "lat")), "WGS84")

distance_from_pb(x = california, ep = na_pa, pb = san_andreas, tangential = TRUE)
#>    [1] 2.502915e+00 4.175856e+00 2.728180e+00 3.523382e+00 5.998119e+00
#>    [6] 6.988708e+00 6.692391e+00 4.979601e+00 5.574181e+00 5.174488e+00
#>   [11] 6.119393e+00 3.432135e+00 3.586663e+00 1.171134e+00 3.305688e+00
#>   [16] 1.353990e+00 5.866841e-01 2.234086e+01 2.813277e+01 2.333412e+01
#>   [21] 2.421695e+01 2.351267e+01 2.071686e+01 1.725346e-02 1.703932e+01
#>   [26] 2.848799e+01 1.973766e+01 1.823931e-01 1.972723e+01 1.909895e+01
#>   [31] 1.703374e+01 2.126209e-02 4.447116e+00 1.710576e+01 1.655899e+01
#>   [36] 2.832636e-01 2.832636e-01 3.635603e-01 1.966582e+00 2.088483e+00
#>   [41] 1.793655e+00 1.620363e+01 4.968503e+00 1.129094e+00 5.489660e-01
#>   [46] 3.164810e-01 3.164810e-01 6.063305e-01 1.438418e+00 1.663966e+00
#>   [51] 2.034735e+01 2.034245e+01 1.595832e+01 3.153326e-01 2.252963e-01
#>   [56] 2.825321e+01 2.584352e-01 1.936184e-01 1.936184e-01 1.703058e+01
#>   [61] 8.957626e-02 4.544694e-01 1.654612e+01 1.584452e+01 2.028649e+01
#>   [66] 2.875482e+01 2.875482e+01 4.762046e-01 1.710583e+01 1.704983e+01
#>   [71] 1.191793e-01 1.696241e+01 5.622517e-02 1.704648e+01 1.904321e+01
#>   [76] 1.593860e+01 1.703234e+01 1.426976e-02 4.811899e-01 1.650673e+01
#>   [81] 1.706031e+01 1.906427e+01 1.765262e+01 3.635603e-01 3.321457e+00
#>   [86] 1.966582e+00 1.540193e+00 2.088483e+00 1.738564e+00 2.538034e+01
#>   [91] 1.365914e-01 2.038894e+00 1.793655e+00 7.310411e-01 4.449104e-01
#>   [96] 1.620363e+01 7.264238e-01 1.755532e+00 1.129094e+00 4.152987e-01
#>  [101] 5.379855e-01 5.489660e-01 7.100832e-01 1.076948e+00 6.063305e-01
#>  [106] 1.368707e+00 1.440551e+00 8.803958e-01 1.663966e+00 1.595832e+01
#>  [111] 7.961964e-01 3.153326e-01 5.796875e-02 1.661256e+00 1.711223e+00
#>  [116] 1.623850e+01 1.954145e-01 2.252963e-01 3.073768e-01 2.825321e+01
#>  [121] 5.279303e-01 9.408314e-01 8.831377e-01 8.123143e-01 6.409156e-01
#>  [126] 1.094626e+00 1.514270e+00 8.049759e-01 2.824563e+01 4.770073e-01
#>  [131] 4.805097e-01 4.041367e-01 2.846232e+01 2.797155e-01 2.584352e-01
#>  [136] 2.284436e-01 1.658544e-01 1.686027e+01 2.445293e-01 1.882993e+01
#>  [141] 3.724949e-01 1.654508e+01 1.710776e+01 1.388087e+00 1.704648e+01
#>  [146] 3.227531e-01 2.905426e+01 8.957626e-02 1.756547e+01 4.544694e-01
#>  [151] 1.677694e+01 1.678964e+01 1.671024e+01 1.649939e+01 2.229163e+01
#>  [156] 1.912100e+01 1.937088e+01 2.872904e+01 1.986525e+01 3.178409e-01
#>  [161] 1.935881e+01 1.957386e+01 3.268638e-01 3.435121e-01 1.988440e+01
#>  [166] 3.118803e-01 1.108753e+00 2.906984e+01 2.899695e+01 2.899514e+01
#>  [171] 8.398762e-02 4.762046e-01 2.421422e+01 2.848881e+00 2.767384e+01
#>  [176] 2.839865e+01 2.273505e+01 1.995755e-01 2.306135e+01 1.777305e+01
#>  [181] 2.068266e+01 2.068266e+01 1.113941e+00 1.851124e+00 1.786885e+00
#>  [186] 2.064671e+01 1.785475e+00 9.892743e-01 7.982335e-01 2.917501e+01
#>  [191] 2.923401e+01 2.850139e+01 1.738511e+00 4.063835e-01 3.753758e+00
#>  [196] 2.764840e+00 2.776384e+01 3.641020e-01 1.566769e+00 5.842877e+00
#>  [201] 2.594225e+01 2.697199e+00 4.178237e+00 4.054729e+00 4.259563e+00
#>  [206] 2.735593e+00 5.807505e+00 5.704827e+00 4.646306e+00 1.346606e+00
#>  [211] 2.703605e+01 1.027366e+00 1.178019e+00 1.094801e+00 2.814619e+01
#>  [216] 3.315708e-01 2.224725e-01 3.452873e-01 3.977952e-01 4.178601e-01
#>  [221] 4.231232e-01 3.025293e-01 3.189590e-01 2.761470e-01 2.416980e-01
#>  [226] 4.858285e-01 5.423321e-01 2.455524e-01 1.573707e-01 1.037705e-01
#>  [231] 1.012371e-01 1.008707e+00 2.783571e+01 7.146348e-01 9.827882e-01
#>  [236] 1.525088e+00 2.043347e+00 2.462774e+00 2.993302e+00 3.368542e+00
#>  [241] 5.422258e-01 1.328603e+00 2.731242e+01 4.959944e-01 9.355166e-03
#>  [246] 2.448882e-01 4.484297e+00 4.244096e+00 4.827607e+00 4.337438e+00
#>  [251] 2.362608e+00 3.446565e+00 1.145743e+00 1.107185e+00 1.054885e+00
#>  [256] 2.239069e-01 1.946865e-01 3.691997e-01 4.315749e-01 3.992479e-01
#>  [261] 4.170775e-01 1.901228e-01 3.423125e-01 2.196386e-01 2.957192e-01
#>  [266] 1.079356e+00 1.734660e-01 2.654727e-01 8.170978e-03 4.309689e-01
#>  [271] 1.731148e+00 6.534366e+00 1.706513e+01 1.339996e+00 2.320448e+00
#>  [276] 4.954653e+00 4.160636e+00 1.365192e+00 6.856576e+00 2.064182e+01
#>  [281] 5.310412e+00 2.266667e+00 4.502583e+00 2.897814e+00 3.431883e+00
#>  [286] 3.181044e+00 3.229962e+00 3.440581e+00 6.384704e+00 7.163255e+00
#>  [291] 5.690663e+00 5.792145e+00 5.071525e+00 6.095582e+00 7.169030e+00
#>  [296] 4.482796e+00 2.295024e+01 1.935678e+01 2.119200e+01 2.071686e+01
#>  [301] 2.359975e+01 1.749711e+01 1.360286e+00 2.183532e-01 7.129157e-01
#>  [306] 1.141717e-01 2.641181e+01 5.077925e-02 1.010947e-01 1.158870e+00
#>  [311] 1.208642e+00 1.561839e+01 2.472098e+01 2.958213e-02 2.021920e+01
#>  [316] 2.808682e+00 2.312146e+00 2.286429e+00 1.962605e+00 2.176087e+01
#>  [321] 1.796795e+00 2.330032e+01 1.747232e+00 2.228168e+01 2.404564e+01
#>  [326] 1.581377e+00 2.185219e+00 1.023055e+00 2.029097e+01 1.761694e-02
#>  [331] 3.908482e+00 5.625622e+00 5.147360e+00 4.207851e+00 4.889130e+00
#>  [336] 4.564068e+00 5.285841e+00 4.442558e+00 3.957337e+00 4.122286e+00
#>  [341] 3.916125e+00 3.243609e+00 1.950969e+00 2.972957e+00 9.218946e-01
#>  [346] 1.855356e+01 2.008617e+01 1.784963e+01 1.590433e+01 6.099315e-01
#>  [351] 6.099315e-01 6.293326e-01 3.753758e+00 5.970838e-01 3.685267e+00
#>  [356] 1.662000e-01 4.341607e-01 4.341607e-01 5.117069e-01 1.068938e+00
#>  [361] 5.605650e-01 5.605650e-01 6.702395e-01 1.908671e+00 6.391400e-01
#>  [366] 1.721063e+00 1.921489e+00 2.305969e+00 1.263068e+00 1.416490e+00
#>  [371] 1.432060e+00 1.428040e+00 1.003300e+00 1.847471e+00 5.742281e+00
#>  [376] 6.835879e+00 6.100493e+00 1.762803e+01 6.138001e-02 2.455524e-01
#>  [381] 2.799247e-01 1.384849e+01 2.228006e+01 3.717122e+00 3.015437e+00
#>  [386] 1.547597e-01 6.975717e+00 6.219969e+00 2.839865e+01 5.271472e-01
#>  [391] 5.924182e+00 3.715359e+00 6.336617e+00 3.006271e+01 1.009709e+00
#>  [396] 2.465823e+01 4.230566e-01 6.105071e-01 5.871438e-02 1.778591e-01
#>  [401] 1.247038e+00 2.462702e+01 2.195690e+01 2.462774e+01 1.349269e+00
#>  [406] 1.980135e+00 2.522702e+00 1.892293e+00 2.546805e-01 2.678963e+01
#>  [411] 3.530260e-02 2.090959e+01 1.174136e+00 2.068600e+00 7.498093e-01
#>  [416] 6.103272e-03 6.649827e-01 3.849684e-01 2.850030e+01 2.848095e+01
#>  [421] 2.816890e+01 2.814948e+01 2.812261e+01 9.792338e-03 4.477683e-02
#>  [426] 7.199217e-02 7.068642e-02 2.411566e+01 7.527075e-01 3.383173e+00
#>  [431] 1.279128e+00 1.810120e-01 5.625327e-02 2.164906e-01 6.724565e-01
#>  [436] 4.221540e-01 3.060536e-01 7.957020e-02 5.366958e-01 2.637473e+01
#>  [441] 1.503860e-01 2.085455e-01 1.456053e-01 1.930575e-01 3.735918e-03
#>  [446] 1.508306e-01 1.860093e-01 1.213923e+00 2.919142e-01 3.408766e-01
#>  [451] 1.189856e-01 3.139888e-01 7.622469e-02 1.228253e-01 3.992579e-01
#>  [456] 8.721751e-03 6.530113e-03 1.653334e-01 1.211416e-01 1.517264e-01
#>  [461] 7.884080e-02 1.789686e-01 1.099045e-01 5.088971e-02 5.046602e-02
#>  [466] 1.203874e-02 1.604923e-02 4.984886e-01 5.141800e-01 4.594769e-01
#>  [471] 4.809571e-01 4.798707e-01 3.421817e-01 3.582150e-01 3.618906e-01
#>  [476] 8.809152e-01 3.187230e-01 2.637859e-01 2.795348e-01 5.364313e-01
#>  [481] 5.051731e-01 2.839285e-01 6.694328e-02 2.245478e-01 7.174486e-01
#>  [486] 7.812708e-01 1.998334e-02 4.305803e-03 1.105434e-02 1.694945e-01
#>  [491] 5.667691e-01 6.304408e-01 1.073504e-01 9.200225e-02 7.564509e-02
#>  [496] 6.027811e-02 4.423029e-02 5.230635e-01 1.470741e-01 1.002522e-01
#>  [501] 4.678946e-01 4.840659e-01 2.225626e-01 1.911715e-01 9.612270e-02
#>  [506] 4.241145e-01 4.878482e-01 5.523924e-01 1.871483e-01 1.399217e-01
#>  [511] 4.285109e-01 4.447082e-01 9.428625e-01 5.616634e-01 3.226251e-01
#>  [516] 3.072568e-01 2.912029e-01 2.748124e-01 2.433354e-01 2.275919e-01
#>  [521] 2.114932e-01 3.411638e-01 3.570217e-01 3.732271e-01 3.887731e-01
#>  [526] 4.053274e-01 3.982009e-01 2.073621e-01 4.094995e-01 4.416899e-01
#>  [531] 4.739170e-01 3.950137e-01 3.789609e-01 3.471780e-01 3.310984e-01
#>  [536] 3.156860e-01 2.992599e-01 2.838286e-01 3.016559e-01 3.172086e-01
#>  [541] 1.119022e+00 4.992343e-01 4.674923e-01 4.514317e-01 4.350342e-01
#>  [546] 4.196339e-01 1.008271e-01 1.173161e-01 2.621251e-01 2.783731e-01
#>  [551] 4.138099e-02 1.377231e-01 5.400604e-01 5.239920e-01 5.075866e-01
#>  [556] 6.101476e-02 7.719960e-02 2.388453e-01 6.597462e-01 7.251048e-01
#>  [561] 5.802281e-01 5.648148e-01 1.086885e-02 4.978307e-03 2.118007e-02
#>  [566] 3.739058e-02 1.663716e-01 1.037907e-01 9.007909e-02 6.214280e-01
#>  [571] 9.970895e-02 6.766351e-02 3.523541e-02 3.116113e-03 1.312922e-02
#>  [576] 9.447937e-02 4.952158e-01 5.108513e-01 1.562556e-01 1.397167e-01
#>  [581] 1.241775e-01 5.888803e-02 4.297019e-02 2.669925e-02 4.225694e-01
#>  [586] 4.392177e-01 4.718749e-01 4.878785e-01 7.706144e-01 2.235086e-02
#>  [591] 4.931106e-01 5.252026e-01 5.583355e-01 1.969904e-01 1.804251e-01
#>  [596] 1.641822e-01 1.479304e-01 1.313379e-01 9.946656e-02 8.284653e-02
#>  [601] 6.655002e-02 3.828212e-01 4.322042e-01 4.478995e-01 2.736455e-01
#>  [606] 4.475816e-01 2.533190e-01 2.370685e-01 2.204772e-01 2.042088e-01
#>  [611] 1.879316e-01 1.719909e-01 2.450468e-01 2.613599e-01 2.824179e-01
#>  [616] 3.151408e-01 2.684970e+01 2.688290e+01 5.213362e-01 5.048564e-01
#>  [621] 4.886972e-01 4.725291e-01 2.934444e-01 2.771684e-01 1.721165e-01
#>  [626] 2.051001e-01 2.214386e-01 5.625256e-01 5.460197e-01 1.320964e-01
#>  [631] 1.487762e-01 1.651314e-01 1.814954e-01 3.773316e-02 2.026598e-01
#>  [636] 4.265944e-02 1.244626e-01 1.575859e-01 5.174694e-02 2.550320e+01
#>  [641] 6.804321e-02 8.444955e-02 4.262429e-02 9.542095e-03 2.652576e+01
#>  [646] 2.646416e+01 2.648059e+01 2.701440e+01 2.703086e+01 2.694035e+01
#>  [651] 2.576999e+01 2.578680e+01 9.286047e-01 2.793055e-01 2.430154e+01
#>  [656] 2.425620e+01 2.761502e+01 1.442882e-01 4.619611e-01 2.610686e+01
#>  [661] 5.348034e-02 7.000288e-02 2.391485e+01 2.393173e+01 1.170003e+00
#>  [666] 1.136465e+00 1.103586e+00 2.743359e-01 2.919390e-01 7.115367e-01
#>  [671] 2.346334e+01 1.323679e+00 1.224384e+00 1.207758e+00 8.070182e-01
#>  [676] 2.338357e+01 2.340052e+01 1.320109e+00 1.287570e+00 8.032377e-01
#>  [681] 7.030191e-01 2.375497e+01 1.547112e-01 1.885834e-01 1.382590e+00
#>  [686] 1.366335e+00 8.987590e-01 1.409014e+00 1.392025e+00 9.909566e-01
#>  [691] 9.746493e-01 9.576347e-01 9.409762e-01 9.246415e-01 9.076008e-01
#>  [696] 8.909143e-01 2.326911e+01 2.328609e+01 2.287364e+01 4.763574e-01
#>  [701] 1.049928e+00 1.033605e+00 1.016573e+00 9.828495e-01 9.497870e-01
#>  [706] 1.530658e+00 2.555763e+01 2.502867e+01 2.309751e+01 2.316610e+01
#>  [711] 1.126325e+00 1.108961e+00 1.092622e+00 1.071507e+00 2.824859e+01
#>  [716] 1.151702e+00 1.117930e+00 1.716638e+00 1.223532e+00 1.189693e+00
#>  [721] 2.528003e+01 2.589169e+01 2.472133e+01 1.316954e+00 2.314537e+01
#>  [726] 5.104820e-02 2.797264e+01 2.406530e+01 2.663881e+01 2.853222e+01
#>  [731] 2.848629e+01 8.913862e-01 3.957239e-01 2.091815e+00 2.642209e+01
#>  [736] 1.031749e+00 7.336709e-01 6.994488e-01 2.116804e+01 2.373793e+01
#>  [741] 2.204228e+01 3.202356e-01 2.824282e+01 2.104083e+01 2.270736e+01
#>  [746] 2.272462e+01 2.274223e+01 8.742510e-01 8.576664e-01 2.234403e+01
#>  [751] 2.236099e+01 2.237832e+01 2.239566e+01 2.264333e+01 1.984381e+01
#>  [756] 2.225007e+01 2.226739e+01 2.228436e+01 2.231904e+01 2.188487e+01
#>  [761] 2.222498e+01 2.185961e+01 2.029338e+01 1.968667e+01 1.970428e+01
#>  [766] 2.179568e+01 5.667724e-01 2.024683e+01 2.729118e+01 1.962840e+01
#>  [771] 1.924325e+01 2.166639e+01 2.786862e+01 2.011849e+01 2.138931e+01
#>  [776] 2.140678e+01 1.914932e+01 1.916730e+01 1.918495e+01 1.920260e+01
#>  [781] 2.890384e+01 2.713251e+01 2.134232e+01 2.777431e+01 2.134740e+01
#>  [786] 2.772765e+01 1.909127e+01 1.910893e+01 2.880982e+01 1.562288e+01
#>  [791] 1.563987e+01 1.866418e+01 1.554610e+01 1.857041e+01 3.594439e-01
#>  [796] 1.645482e+01 2.868628e+01 1.636096e+01 1.810318e+01 1.812090e+01
#>  [801] 1.918707e+01 1.883818e+01 2.859238e+01 1.962226e+01 1.513843e+01
#>  [806] 1.779009e+01 4.138216e-01 1.655622e+01 1.720861e+01 2.078075e+01
#>  [811] 1.391259e-02 2.896000e+01 9.407148e-01 1.093043e+00 5.966800e-01
#>  [816] 1.932585e+01 1.993637e+01 1.576284e+01 1.841307e+01 1.771718e+01
#>  [821] 7.914576e-01 5.575841e-01 2.953083e+01 1.290666e-01 3.758432e-01
#>  [826] 1.332577e-01 1.047103e-01 7.624761e-02 2.361882e+00 2.739110e+00
#>  [831] 8.997579e-01 2.064522e-01 2.210400e+01 2.352924e-01 2.235997e+00
#>  [836] 2.883377e+00 4.150785e-02 7.102411e-01 1.144411e+00 1.905166e-01
#>  [841] 2.902864e+01 2.446637e+00 3.039357e+01 5.866132e-01 4.342825e-01
#>  [846] 1.200735e+00 1.708476e+01 1.888802e+01 9.585510e-01 4.558873e-01
#>  [851] 2.319253e-01 2.765721e-01 4.000853e-01 2.713585e-01 3.416195e-01
#>  [856] 3.589653e-01 2.319281e-01 2.295995e-01 5.943383e-03 8.093632e-02
#>  [861] 2.302056e-01 1.836780e-01 9.238666e-03 6.103174e-02 3.573937e-02
#>  [866] 2.934883e-01 1.061572e+00 1.346936e-01 6.230740e-02 2.867402e-02
#>  [871] 7.371402e-01 8.093632e-02 6.971811e-02 4.926570e-02 3.756732e+00
#>  [876] 3.771193e+00 8.642116e-02 8.953412e-02 2.431604e-02 1.717954e+00
#>  [881] 6.909710e-02 9.878303e-01 8.096464e-01 4.687944e-01 2.477995e+01
#>  [886] 5.331754e+00 5.326572e+00 6.788709e+00 5.051414e+00 6.097319e+00
#>  [891] 6.863096e+00 9.743709e-02 4.071807e-01 2.967433e-03 1.121212e-01
#>  [896] 2.504893e-01 1.586968e+01 1.902116e-01 1.079771e+00 7.533692e-01
#>  [901] 2.814404e+01 1.116193e-01 2.800449e-01 8.477329e-01 1.324659e+00
#>  [906] 1.417612e+00 2.833891e+00 2.495310e+00 2.045050e+00 1.094036e+00
#>  [911] 2.496492e-01 3.416197e-01 3.234177e+00 4.229977e+00 3.414542e+00
#>  [916] 1.600925e+00 1.047743e+00 2.132921e+01 4.992456e-01 2.052656e-01
#>  [921] 1.567629e-01 1.136578e+00 2.143637e-03 1.029656e-02 2.025342e-02
#>  [926] 4.012383e-02 3.602121e-01 1.855959e-01 1.714103e-01 2.039939e-01
#>  [931] 2.394422e-04 5.242471e-01 1.081390e-01 1.010762e-01 6.616744e-02
#>  [936] 1.253315e-01 2.042093e-01 4.821727e-02 2.901786e-01 4.048676e-01
#>  [941] 5.133380e-01 3.848128e-01 4.814185e-01 5.108279e-01 5.189440e-01
#>  [946] 5.270625e-02 3.031030e-01 5.249987e-01 7.405038e-02 5.407493e-02
#>  [951] 2.944132e-01 2.168790e-01 1.628242e-01 1.411883e-01 2.911484e-01
#>  [956] 2.748988e-01 1.379469e-01 1.155033e+00 2.596922e+00 3.938875e-01
#>  [961] 2.207803e+01 6.761387e-02 1.139371e+00 2.285489e+01 5.513856e-02
#>  [966] 2.101003e-01 1.019152e+00 2.124236e-01 8.077424e-01 6.139480e-02
#>  [971] 1.588841e-01 1.181694e-01 2.018158e-01 6.884736e-02 2.212501e-01
#>  [976] 1.917895e-01 1.256564e-01 4.915056e-01 1.606028e+00 1.606028e+00
#>  [981] 1.606028e+00 8.127499e-02 7.023030e-01 5.932080e-02 9.055378e-02
#>  [986] 1.211092e-01 3.691053e-01 6.391173e-01 6.292449e-01 1.472693e-01
#>  [991] 4.686161e-01 1.569485e-01 6.259650e-01 6.259650e-01 6.686515e-01
#>  [996] 5.318900e-02 5.437098e-01 7.947827e-01 3.204032e-01 4.599514e-01
#> [1001] 1.130869e+00 1.436316e-01 8.763811e-01 1.590228e-01 9.364139e-01
#> [1006] 5.315193e-01 2.187204e-02 9.907233e-01 1.007137e+00 1.000983e+00
#> [1011] 2.818094e-01 1.284469e-01 4.340060e-01 1.984422e-01 6.252272e+00
#> [1016] 2.231472e-01 2.187741e-01 3.602568e-01 1.560225e-01 1.028736e-01
#> [1021] 1.532753e-01 3.917609e-01 2.890877e+00 2.807210e-02 3.922551e-02
#> [1026] 2.769350e-02 3.387214e-02 4.791112e-02 2.442886e-02 2.772772e+01
#> [1031] 1.899424e+01 9.191681e-01 5.124304e-01 9.168834e-02 2.664447e-01
#> [1036] 4.621136e-01 2.301060e-01 4.114603e-02 6.423796e-01 5.072419e-03
#> [1041] 5.277108e-01 5.953869e-01 5.335905e-01 2.313491e-01 2.616166e-01
#> [1046] 2.243047e-02 3.738171e+00 5.812773e-02 7.023351e-01 7.062460e+00
#> [1051] 8.072404e-01 1.906651e-01 4.002256e-02 9.845915e-02 2.678769e+00
#> [1056] 1.373698e+00 1.003306e+00 7.123186e-01 2.679762e-01 1.818037e+00
#> [1061] 1.181964e+00 7.093549e+00 2.390561e+01 7.274360e-01 3.320708e-01
#> [1066] 7.739420e-02 2.821816e+01 3.178125e+00 3.057402e+01 5.211843e-02
#> [1071] 1.356702e-02 6.099358e-02 1.182690e+00 7.586436e-01 7.618500e-01
#> [1076] 1.361267e-01 3.543238e-01 1.835537e-01 2.704647e-01 4.658486e+00
#> [1081] 9.049038e-01 1.801296e+00 1.663850e+01 1.743182e+01 1.866721e+01
#> [1086] 1.359863e+00 2.053034e+01 2.788580e+01 1.137194e+00 2.498012e+00
#> [1091] 1.312108e-01 3.601051e+00 3.601051e+00 5.381526e+00 1.174194e+00
#> [1096] 1.162353e+00 1.230573e+00 1.116530e+00 1.179108e+00 1.125485e+00
#> [1101] 1.233682e+00 1.198711e+00 1.128484e+00 1.158902e+00 1.162442e+00
#> [1106] 1.151829e+00 1.127992e+00 1.163223e+00 1.170409e+00 1.178026e+00
#> [1111] 1.184714e+00 1.166035e+00 1.191892e+00 1.237278e+00 1.139956e+00
#> [1116] 1.140866e+00 1.080980e+00 1.110516e+00 1.143713e+00 1.181825e+00
#> [1121] 1.156121e+00 1.132327e+00 1.120283e+00 1.146867e+00 1.187138e+00
#> [1126] 1.113138e+00 1.082470e+00 1.082470e+00 1.142369e+00 1.139062e+00
#> [1131] 1.169497e+00 1.163488e+00 1.087807e+00 1.007053e+00 1.067432e+00
#> [1136] 1.070568e+00 1.070568e+00 1.070586e+00 1.070586e+00 1.007641e+00
#> [1141] 1.007641e+00 1.172177e+00