Draws small-circles around Euler pole

Construct a line that has a constant distance to the Euler pole.

Usage

eulerpole_smallcircles(x, sm, small_circle = NULL)

Arguments

x

data.frame containing coordinates of Euler pole in lat, lon, and rotation angle (optional)

sm

numeric, angle between small circles (in degree); default: 10

Value

An object of class SpatialLinesDataFrame

Details

if angle is given: output additionally gives absolute velocity on small-circle (degree/Myr -> km/Myr)

Author

Tobias Stephan

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
euler$angle <- euler$rate
# Pacific plate
eulerpole_smallcircles(euler)
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#> Warning: implicit list embedding of S4 objects is deprecated
#>                                      geometry small_circle.small_circle
#> 0      MULTILINESTRING((101.833 -48.709 ...))                         0
#> 10  MULTILINESTRING((115.9334 -45.76261 ...))                        10
#> 20  MULTILINESTRING((128.2862 -41.28265 ...))                        20
#> 30  MULTILINESTRING((138.8716 -35.66505 ...))                        30
#> 40  MULTILINESTRING((147.9749 -29.24873 ...))                        40
#> 50  MULTILINESTRING((155.9604 -22.29108 ...))                        50
#> 60  MULTILINESTRING((163.1686 -14.98091 ...))                        60
#> 70  MULTILINESTRING((169.8928 -7.459644 ...))                        70
#> 80  MULTILINESTRING((176.3874 0.1596386 ...))                        80
#> 90  MULTILINESTRING((-177.1131 7.776871 ...))                        90
#> 100  MULTILINESTRING((-170.374 15.29171 ...))                       100
#> 110 MULTILINESTRING((-163.1395 22.59024 ...))                       110
#> 120 MULTILINESTRING((-155.1144 29.52944 ...))                       120
#> 130 MULTILINESTRING((-145.9563 35.91806 ...))                       130
#> 140 MULTILINESTRING((-135.3016 41.49543 ...))                       140
#> 150 MULTILINESTRING((-122.8721 45.91936 ...))                       150
#> 160  MULTILINESTRING((-108.707 48.79309 ...))                       160
#> 170 MULTILINESTRING((-93.42389 49.76931 ...))                       170
#> 180     MULTILINESTRING((-78.167 48.709 ...))                       180
#>     small_circle.abs_vel
#> 0                0.00000
#> 10              15.11688
#> 20              29.77444
#> 30              43.52731
#> 40              55.95764
#> 50              66.68772
#> 60              75.39152
#> 70              81.80459
#> 80              85.73207
#> 90              87.05463
#> 100             85.73207
#> 110             81.80459
#> 120             75.39152
#> 130             66.68772
#> 140             55.95764
#> 150             43.52731
#> 160             29.77444
#> 170             15.11688
#> 180              0.00000